#207 - update references to how people can talk about the calendar
Stikeout references to the mailing list and add a pointer to the
GitHub repo. This retconning makes the text seem a bit weird in
places, but I think changing things too much is also weird since we
changing the historical record.

This also updates all the GitHub pointer to go to perl-advent/Perl-Advent
instead of any other repo used in the past.