Run "Deploy" workflow only on main repository
I'd noticed that my fork of the repository kept trying to build and
deploy the site by running the "Deploy" GitHub workflow/action.  This
action keeps failing because (quite rightly) I don't have the
permissions to deploy the site and forks should be trying to do that.

The change implemented here checks that the repository owner matches
that of the main repository and hence the action only runs on that
repository and no longer on forks, which I think should be the correct
behaviour.

The inspiration for this change came from the GitHub community
discussion [Have github action only run on master repo, and not on
forks?](https://github.com/orgs/community/discussions/26409).